incorporate review comments
- README.adoc
Clarify that the wavelength of the IR LED should be compatible with the
detector used;
Correct the number of jumper wires in the BOM.

- Wiring diagram
Use red wire for +3v3 and black for GND.

- Fritzing source file
Fix having added the wrong file.

- ir_loopback/ir_loopback.c
Correct typo in comment.

- nec_receive_library/nec_receive.pio
Amend comment to clarify that the provided 'init' function sets the
state machine clock divider relative to the system clock.
Amend comment to clarify that the 'init' function sets up the ISR as
described.