Consider filter drop for Vref and use max ADC as 2^12 - 1
Pico datasheet, 4.3. Using the ADC: "The ADC draws current (about 150μA if the temperature sense diode is disabled, but it varies from chip to chip) and therefore there will be an inherent offset of about 150μA*200 = ~30mV. There is a small difference in current draw when the ADC is sampling (about +20μA)"
Thus when sampling Vref is nominal (150+20)μA*200 = ~34mV

Splitting the variable in two lines makes the change for the clamped situation more intuitive. 

The value range of a 12 bit resolution is 0-4095, i.e. 2^12 minus one.