Remove implicit conversions between function and data pointers
IAR generates errors Pe144 'a value of type "int32_t (*)(int32_t)" cannot
be used to initialize an entity of type "void *" and Pe513 'a value of
type "int32_t (*)(int32_t)" cannot be assigned to an entity of type
"void *"'.

Also, when citing the name of a specific function, there's no need to use
the '&' operator before the symbol - it's already typed as a function
pointer.