Some example improvements (#634)
* Use async context for led in iperf example

Improve led handling by using the async context to turn the led on
and off.
Fix assert in poll mode when cyw43_arch_disable_sta_mode is called in
interrupt context.

* Improve wifi scan example

Use async context to initiate a scan in future instead of keeping track
of time in the main loop.
Detect a key press to exit the example.

* Improve NTP example

Use async context NTP requests, rather than keeping track of time in the
main loop.
Detect a key press to exit the example.