Properly map IO and PAD register offsets for QSPI
In the UART bootloader example the QSPI pins are used for normal IO and
a `qspi_gpio_set_function()` function is used locally to setup the pins.
This function didn't take into account that the IO and PAD register
banks for the QSPI IOs are not in the same order. We correct this by
adding a lookup table to map between the two.